Tidy up the mock server env, update documentation.
