+++ src/package.json
@@ -22,12 +22,13 @@
   "dev": "cross-env NODE_ENV=development npm-run-all copy-assets tsoa-build ts-build ks-dev",
   "ks-dev": "cross-env NODE_ENV=development DISABLE_LOGGING=true keystone dev --entry=dist/server.js",
   "dev2": "cross-env NODE_ENV=development DISABLE_LOGGING=true keystone --entry=dist/index.js",
+  "mock-server": "nodemon ./test/mock-server/server.js",
   "ks-build": "cross-env NODE_ENV=production keystone build --entry=dist/server.js --out=dist2",
   "start": "cross-env NODE_ENV=development keystone start --entry=dist/server.js",
   "build": "npm-run-all copy-assets tsoa-build ts-build ks-build copy-keystone-admin-assets",
   "create-tables": "cross-env CREATE_TABLES=true keystone create-tables --entry=dist/server.js",
   "lint": "eslint .",
-  "lint-ts": "eslint ./nextapp --ext .ts,.tsx --quiet & tsc -p ./nextapp/tsconfig.json --noEmit",
+  "lint:ts": "eslint ./nextapp --ext .ts,.tsx --quiet & tsc -p ./nextapp/tsconfig.json --noEmit",
   "test": "cross-env NODE_ENV=test jest --config ./jest.config.js --coverage --detectOpenHandles",
   "test:watch": "jest -w --coverage",
   "test:node": "node_modules/.bin/mocha",
